refactor: extract hooks and sections
